Refactor to project references - renamed "compile" script to "build". called it in "pretest" and used it "watch". - simplified travis config. `sudo: false` is no longer needed. test now always runs build, so no need to call it explicitly. - upgrade to latest typescript (3.9.3) to get project references working as expected. - upgrade to latest mocha to fix all audit issues. - added "files" field to `package.json`, so that tests won't get published. - enabled declaration maps and ensured src folders are also published for proper ctrl+click in consumer libraries.
