+++ .travis.yml
@@ -1,5 +1,4 @@
 language: node_js
-sudo: false
 
 node_js:
   - 10
@@ -10,5 +9,4 @@ install:
   - npm install
 
 script:
-  - npm run compile
   - npm test
