+++ browser/main-browser.ts
@@ -1,6 +1,6 @@
-import { setPseudo, localize, Options } from './common';
+import { setPseudo, localize, Options } from '../common/common';
 
-export { MessageFormat, BundleFormat, Options, LocalizeInfo, LocalizeFunc, LoadFunc, KeyInfo } from './common';
+export { MessageFormat, BundleFormat, Options, LocalizeInfo, LocalizeFunc, LoadFunc, KeyInfo } from '../common/common';
 
 export function loadMessageBundle(_file?: string) {
 	return localize;
