+++ package.json
@@ -11,24 +11,33 @@
 	"bugs": {
 		"url": "https://github.com/Microsoft/vscode-nls/issues"
 	},
-	"main": "./lib/main.js",
-	"typings": "./lib/main.d.ts",
-	"browser": "./lib/main-browser.js",
+	"main": "./lib/node/main-node.js",
+	"typings": "./lib/node/main-node.d.ts",
+	"browser": "./lib/browser/main-browser.js",
 	"devDependencies": {
-		"@types/node": "^12.12.29",
 		"@types/mocha": "^7.0.2",
-		"mocha": "^7.1.0",
-		"typescript": "^3.8.3",
-		"eslint": "^6.8.0",
+		"@types/node": "^12.12.29",
 		"@typescript-eslint/parser": "^2.24.0",
-		"rimraf": "^3.0.2"
+		"eslint": "^6.8.0",
+		"mocha": "^7.2.0",
+		"rimraf": "^3.0.2",
+		"typescript": "^3.9.3"
 	},
 	"scripts": {
 		"prepublishOnly": "npm run lint && npm run compile && npm run test",
-		"compile": "tsc -p ./src",
-		"watch": "tsc -w -p ./src",
+		"build": "tsc --build tsconfig.json",
+		"watch": "npm run build -- -w",
 		"clean": "rimraf lib",
+		"pretest": "npm run build",
 		"test": "mocha",
-		"lint": "eslint --config .eslintrc.json ./src/**/*.ts"
-	}
+		"lint": "eslint --config .eslintrc.json ./**/*.ts"
+	},
+	"files": [
+		"common",
+		"browser",
+		"node",
+		"lib/common",
+		"lib/browser",
+		"lib/node"
+	]
 }
