--- src/tsconfig.json
@@ -1,13 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "es5",
-		"lib": [ "es2015", "es2015.promise" ],
-		"module": "commonjs",
-		"moduleResolution": "node",
-		"sourceMap": true,
-		"inlineSources": false,
-		"declaration": true,
-		"stripInternal": true,
-		"outDir": "../lib"
-	}
-}
