+++ tsconfig.base.json
@@ -1,0 +1,14 @@
+{
+	"compilerOptions": {
+		"target": "es5",
+		"lib": [ "es5" ],
+		"module": "commonjs",
+		"moduleResolution": "node",
+		"sourceMap": true,
+		"declaration": true,
+		"declarationMap": true,
+		"stripInternal": true,
+		"composite": true,
+		"types": []
+	}
+}
